pyproject.toml
[tool.poetry]
name = "judybot"
version = "0.1.0"
description = ""
authors = ["ZZ3n <ndsrhkd@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

